chore(state and reducers): clean up state and reducers
remove redundant action types
move the center object in state to the centers object
move the event object in state to the events object
delete client/reducers/singleEventReducer.js
delete client/reducers/singleCenterReducer.js
remove center and event from the root reducer
move case for fetching a single event to client/reducers/eventReducer.js
move case for fetching a single center to client/reducers/centerReducer.js
modify necessary components to take advantage of the new shape of state

[finishes #157770845]
