+++ client/actions/actionTypes.js
@@ -11,9 +11,6 @@ export const ADD_CENTER_LOADING = 'ADD_CENTER_LOADING';
 export const FETCH_CENTERS_SUCCESS = 'FETCH_CENTERS_SUCCESS';
 export const FETCH_CENTERS_FAILURE = 'FETCH_CENTERS_FAILURE';
 export const FETCH_CENTERS_LOADING = 'FETCH_CENTERS_LOADING';
-export const FETCH_EVENTS_IN_CENTER_SUCCESS = 'FETCH_EVENTS_IN_CENTER_SUCCESS';
-export const FETCH_EVENTS_IN_CENTER_FAILURE = 'FETCH_EVENTS_IN_CENTER_FAILURE';
-export const FETCH_EVENTS_IN_CENTER_LOADING = 'FETCH_EVENTS_IN_CENTER_LOADING';
 export const FETCH_SINGLE_CENTER_SUCCESS = 'FETCH_SINGLE_CENTER_SUCCESS';
 export const UPDATE_CENTER_SUCCESS = 'UPDATE_CENTER_SUCCESS';
 export const UPDATE_CENTER_FAILURE = 'UPDATE_CENTER_FAILURE';
@@ -31,4 +28,3 @@ export const UPDATE_EVENT_LOADING = 'UPDATE_EVENT_LOADING';
 export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
 export const SET_COMPONENT_NAME_SUCCESS = 'SET_COMPONENT_NAME_SUCCESS';
 export const DELETE_EVENT_SUCCESS = 'DELETE_EVENT_SUCCESS';
-export const DELETE_EVENT_IN_CENTER_SUCCESS = 'DELETE_EVENT_IN_CENTER_SUCCESS';
